renamed tests so can be run with python -m unittest discover -s ./test/unittests  -p 'test_*.py'
