--- tests/unittests/settings_io_test_roundtrip.ini
@@ -1,24 +1,0 @@
-[app]
-save_dir = C:/Users/bened/OneDrive/foundry_scope/scope_foundry/ScopeFoundry/tests/data
-sample = 
-data_fname_format = {timestamp:%y%m%d_%H%M%S}_{measurement.name}.{ext}
-propose_from_file = 
-
-[mm/measure1]
-activation = False
-run_state = stop_first
-progress = 0.0
-profile = False
-string = 0
-float = 0.0
-choices = 0
-array = [false, false, false]
-
-[hw/hardware1]
-connected = False
-debug_mode = False
-string = 0
-float = 0.0
-choices = 0
-protected_int = 0
-
