+++ tests/unittests/test_analyze_nb.py
